backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
faster-whisper
pyannote.audio
torch
torchaudio
pydub
cloudinary
python-dotenv
websockets
httpx
pytest
pytest-asyncio
celery
redis
ffmpeg-python
requests
